Separate Ublas and VTK converters
