+++ dynamic/wrappers/cell_based/AttractingPlaneBoundaryCondition3_3.cppwg.cpp
@@ -1,13 +1,13 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
-#include "PythonObjectConverters.hpp"
+#include "PythonUblasObjectConverters.hpp"
 #include <set>
 #include <vector>
 #include <string>
 #include <map>
 #include "SmartPointers.hpp"
 #include "UblasIncludes.hpp"
-#include "PythonObjectConverters.hpp"
+#include "PythonUblasObjectConverters.hpp"
 #include "AttractingPlaneBoundaryCondition.hpp"
 
 #include "AttractingPlaneBoundaryCondition3_3.cppwg.hpp"
