+++ dynamic/wrappers/cell_based/PlaneBasedCellKiller2.cppwg.cpp
@@ -6,7 +6,7 @@
 #include <map>
 #include "SmartPointers.hpp"
 #include "UblasIncludes.hpp"
-#include "PythonObjectConverters.hpp"
+#include "PythonUblasObjectConverters.hpp"
 #include "PlaneBasedCellKiller.hpp"
 
 #include "PlaneBasedCellKiller2.cppwg.hpp"
