+++ dynamic/wrappers/cell_based/SphereGeometryBoundaryCondition2.cppwg.cpp
@@ -1,6 +1,6 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
-#include "PythonObjectConverters.hpp"
+#include "PythonUblasObjectConverters.hpp"
 #include <set>
 #include <vector>
 #include <string>
