+++ dynamic/wrappers/visualization/VtkScene2.cppwg.cpp
@@ -10,7 +10,7 @@
 #include <vtkOpenGLRenderer.h>
 #include <vtkUnsignedCharArray.h>
 #include <vtkSmartPointer.h>
-#include "PythonObjectConverters.hpp"
+#include "PythonVtkObjectConverters.hpp"
 
 #include "VtkScene2.cppwg.hpp"
